+++ react/__mocks__/vtex.render-runtime.js
@@ -1,5 +1,13 @@
+import React from 'react'
+
 export function Link({ children }) {
   return children
+}
+
+export function ExtensionPoint(props) {
+  return (
+    <div {...props} />
+  )
 }
 
 export function withRuntimeContext(Comp) {