+++ react/legacy/index.js
@@ -1,7 +1,10 @@
 import PropTypes from 'prop-types'
 import React, { Component } from 'react'
 import { withRuntimeContext } from 'vtex.render-runtime'
-import { ProductName, ProductPrice } from 'vtex.store-components'
+import {
+  ProductName,
+  ProductPrice,
+} from 'vtex.store-components'
 
 import ProductSummaryNormal from './components/ProductSummaryNormal'
 import ProductSummarySmall from './components/ProductSummarySmall'
@@ -41,6 +44,8 @@ class ProductSummary extends Component {
     showBadge: PropTypes.bool,
     /** Set the product description visibility */
     showDescription: PropTypes.bool,
+    /** Set the quantity selector visibility */
+    showQuantitySelector: PropTypes.bool,
     /** Text of selling Price's label */
     labelSellingPrice: PropTypes.string,
     /** Text shown on badge */
@@ -73,6 +78,7 @@ class ProductSummary extends Component {
     showInstallments: true,
     showLabels: true,
     showBadge: true,
+    showQuantitySelector: true,
     showCollections: false,
     showDescription: false,
     displayBuyButton: displayButtonTypes.DISPLAY_ALWAYS.value,
@@ -121,6 +127,7 @@ class ProductSummary extends Component {
       showInstallments,
       labelSellingPrice,
       name: showFieldsProps,
+      showQuantitySelector,
     } = this.props
 
     const imageProps = {
@@ -165,6 +172,7 @@ class ProductSummary extends Component {
         nameProps={nameProps}
         priceProps={priceProps}
         buyButtonProps={buyButtonProps}
+        showQuantitySelector={showQuantitySelector}
       />
     )
   }